scheduler/requirements.txt
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
google-cloud-pubsub==2.22.0
requests==2.32.3
